Added swal with confirm dialogues

+++ js/main.js
@@ -16,11 +16,11 @@ const obtenerDatos = (event) => {
     let dias = parseInt(inputDias.value)
 
     if (nombre.length < 3) {
-        alertas("error", "Error!", "Por favor ingrese un nombre con mas de 3 caracteres.", 'animate__animated animate__headShake')
+        alertas("error", "Error!", "Por favor ingrese un nombre con mas de 3 caracteres.", "animate__animated animate__headShake")
     } else if (isNaN(capital) || capital <= 0) {
-        alertas("error", "Error!", "Por favor ingresa un capital mayor a cero.", 'animate__animated animate__headShake')
+        alertas("error", "Error!", "Por favor ingresa un capital mayor a cero.", "animate__animated animate__headShake")
     } else if (isNaN(dias) || dias <= 0) {
-        alertas("error", "Error!", "Por favor ingresa un numero de dias mayor a cero.", 'animate__animated animate__headShake')
+        alertas("error", "Error!", "Por favor ingresa un numero de dias mayor a cero.", "animate__animated animate__headShake")
     } else {
         agregarHistoriales(nombre, capital, dias) // En DOM.js
     }
@@ -30,12 +30,30 @@ const obtenerDatos = (event) => {
 const borrarArray = () => {
     // Chequeamos que el historial de calculos tenga datos cargados
     if (historialCalculadoraArray.length > 0) {
-        historialCalculadoraArray.length = 0
-        const borrarHistorial = document.querySelectorAll(".delete")
-        borrarHistorial.forEach((element) => {
-            element.remove()
-        })
-        alertas("success", "Hecho!", "Su historial de calculos ha sido borrado", "animate__animated animate__bounceIn")
+        Swal
+            .fire({
+                title: "Esta seguro?",
+                text: "Perderas todos los datos!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Si, eliminarlo!",
+                cancelButtonText: "Cancelar",
+                customClass: {
+                    confirmButton: "btn btn-danger m-3",
+                    cancelButton: "btn btn-secondary m-3"
+                },
+                buttonsStyling: false
+            })
+            .then((result) => {
+                if (result.isConfirmed) {
+                    historialCalculadoraArray.length = 0
+                    const borrarHistorial = document.querySelectorAll(".delete")
+                    borrarHistorial.forEach((element) => {
+                        element.remove()
+                    })
+                    alertas("success", "Hecho!", "Su historial de calculos ha sido borrado", "animate__animated animate__bounceIn")
+                }
+            })
     }
     else {
         alertas("error", "Error!", "El historial de calculos ya se encuentra vacio.", "animate__animated animate__headShake")
@@ -54,12 +72,30 @@ const misConsultas = () => {
 // // Llamamos a la funcion borrarConsultas mediante el boton de Borrar Consultas
 const borrarConsultas = () => {
     if (localStorage.length > 0) {
-        localStorage.clear()
-        alertas("success", "Hecho!", "Su historial de consultas ha sido eliminado.", "animate__animated animate__bounceIn")
-        // Eliminamos todos los hijos de #consultasCliente
-        while (btnConsultas.firstChild) {
-            btnConsultas.firstChild.remove()
-        }
+        Swal
+            .fire({
+                title: "Esta seguro?",
+                text: "Perderas todos los datos!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Si, eliminarlo!",
+                cancelButtonText: "Cancelar",
+                customClass: {
+                    confirmButton: "btn btn-danger m-3",
+                    cancelButton: "btn btn-secondary m-3"
+                },
+                buttonsStyling: false
+            })
+            .then((result) => {
+                if (result.isConfirmed) {
+                    localStorage.clear()
+                    // Eliminamos todos los hijos de #consultasCliente
+                    while (btnConsultas.firstChild) {
+                        btnConsultas.firstChild.remove()
+                    }
+                    alertas("success", "Hecho!", "Su historial de consultas ha sido eliminado.", "animate__animated animate__bounceIn")
+                }
+            })
     } else {
         alertas("error", "Error!", "Usted aun no tiene ninguna consulta archivada.", "animate__animated animate__headShake")
     }
